doc: grep: unify configuration variables definitions
The configuration variables for git-grep are duplicated in
"Documentation/git-grep.txt" and "Documentation/config/grep.txt", which
gqcan make maintenance difficult. The first also contains a definition
that is not present in the latter (grep.fullName), and the latter
received a wording improvement that was not replicated in the former:
see 91028f765 ("grep: clarify what `grep.patternType=default` means",
2021-12-05).

To avoid such problems, unify the information in one file and include it
in the other.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>
Signed-off-by: Junio C Hamano <gitster@pobox.com>